local: Update to new DMABUF-based kernel API
The previous DMABUF-based API was refused upstream as it had several
problems and was very complex.

A lot of that complexity was lifted by making the IIO core a DMABUF
importer instead of an exporter. This means that the IIO core is no more
responsible for creating the DMABUF objects, and to map them in the DMA
space.

This task is now delegated to the "udmabuf" kernel driver, available
since kernel 5.19, which allows userspace programs to create DMABUF
objects from memfd buffers.

The DMABUF interface of the IIO core then simply consists in three
IOCTLs: one to attach a DMABUF to the interface, one to detach it, and
one to request a data transfer.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>